feat: 주문 시 프로모션 적용 기능 추가

+++ src/test/java/store/core/model/PromotionTest.java
@@ -11,8 +11,8 @@ class PromotionTest {
         // given
         // when
         String name = "promotion name";
-        int buy = 2;
-        int get = 1;
+        Long buy = 2L;
+        Long get = 1L;
         LocalDate startDate = LocalDate.of(2024, 11, 1);
         LocalDate endDate = LocalDate.of(2024, 11, 30);
         Promotion promotion = new Promotion(name, buy, get, startDate, endDate);
@@ -28,9 +28,9 @@ class PromotionTest {
     void equals_should_be_pass() {
         // given
         // when
-        Promotion promotion1 = new Promotion("equals", 1, 1, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
-        Promotion promotion2 = new Promotion("equals", 2, 1, LocalDate.of(2024, 11, 10), LocalDate.of(2024, 11, 20));
-        Promotion promotion3 = new Promotion("not_equals", 1, 1, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
+        Promotion promotion1 = new Promotion("equals", 1L, 1L, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
+        Promotion promotion2 = new Promotion("equals", 2L, 1L, LocalDate.of(2024, 11, 10), LocalDate.of(2024, 11, 20));
+        Promotion promotion3 = new Promotion("not_equals", 1L, 1L, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
         // then
         Assertions.assertEquals(promotion1, promotion2);
         Assertions.assertNotEquals(promotion1, promotion3);
@@ -40,7 +40,7 @@ class PromotionTest {
     @Test
     void getPromotionQuantity_should_be_pass() {
         // given
-        Promotion promotion = new Promotion("promotion", 2, 1, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
+        Promotion promotion = new Promotion("promotion", 2L, 1L, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
         // when
         Long quantity = 10L;
         LocalDate date = LocalDate.of(2024, 11, 15);
@@ -52,7 +52,7 @@ class PromotionTest {
     @Test
     void getPromotionQuantity_when_not_between_start_and_end_should_return_negative_one() {
         // given
-        Promotion promotion = new Promotion("promotion", 2, 1, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
+        Promotion promotion = new Promotion("promotion", 2L, 1L, LocalDate.of(2024, 11, 1), LocalDate.of(2024, 11, 30));
         // when
         Long quantity = 10L;
         LocalDate date = LocalDate.of(2024, 12, 1);
